Add Gear Menu framework
Added Gear Menu which appears in while playing level after pressing the
"i" key, currently has 20 empty slots and a space for an equipment
interface that is yet to be added. Updated blaster to work better with
new gear menu, and can be seen in gear menu when picked up.