Add equipping and unequipping items from inventory
Items in inventory can now be equipped to one of the three equipment slots, this can be done by double left-clicking the item, this is also how unequiping items works.  Actual functionality of items is right around the corner, just need to discuss   a few things with Ian before I can edit stats though an example is given in EquipSlot.  Made stat fields in PlayerMovement static to allow for editing later.